Move and rename nomination response model to singular
